+++ src/libb64/cencode.c
@@ -13,7 +13,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.5.3
+  Version: 1.5.4
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -38,7 +38,10 @@
   1.5.1   K Hoang      25/12/2021 Fix bug
   1.5.2   K Hoang      27/12/2021 Fix wrong http status header bug
   1.5.3   K Hoang      27/12/2021 Fix authenticate issue caused by libb64
+  1.5.4   K Hoang      12/01/2022 Fix libb64 fallthrough compile warning
  *****************************************************************************************************************************/
+ 
+#if !(ESP32 || ESP8266)
  
 #include "cencode.h"
 
@@ -58,7 +61,7 @@ char base64_encode_value(char value_in)
   if (value_in > 63)
     return '=';
 
-  return encoding[(int)value_in];
+  return encoding[(unsigned int)value_in];
 }
 
 int base64_encode_block(const char* plaintext_in, int length_in, char* code_out, base64_encodestate* state_in)
@@ -87,6 +90,8 @@ int base64_encode_block(const char* plaintext_in, int length_in, char* code_out,
         result = (fragment & 0x0fc) >> 2;
         *codechar++ = base64_encode_value(result);
         result = (fragment & 0x003) << 4;
+        
+        // fall through
 
       case step_B:
         if (plainchar == plaintextend)
@@ -100,7 +105,9 @@ int base64_encode_block(const char* plaintext_in, int length_in, char* code_out,
         result |= (fragment & 0x0f0) >> 4;
         *codechar++ = base64_encode_value(result);
         result = (fragment & 0x00f) << 2;
-
+        
+        // fall through
+        
       case step_C:
         if (plainchar == plaintextend)
         {
@@ -122,6 +129,8 @@ int base64_encode_block(const char* plaintext_in, int length_in, char* code_out,
           *codechar++ = '\n';
           state_in->stepcount = 0;
         }
+        
+        // fall through
       }
   }
 
@@ -161,3 +170,5 @@ int base64_encode_chars(const char* plaintext_in, int length_in, char* code_out)
 
   return len + base64_encode_blockend((code_out + len), &_state);
 }
+
+#endif